Add a function to get raw stroage from a node
this is a useful helper function.

Namely when we make examples, we put them in the node data structure,
and this allows us easily to extract the storage.

I think in the long term we should redesign this type, but for now I
think this is a useful way to expose the inner storage without too
much effort